auth_service/requirements.txt
Flask==2.1.3
PyJWT==2.1.0
Werkzeug==2.0.3
requests==2.26.0
PyMongo==4.10.1